requirements.txt
opencv-python
pyzbar
streamlit
numpy
pillow
pandas
